Refactor Object.
